backend/requirements.txt

flask==2.0.1
flask-cors==3.0.10
numpy==1.19.5
tensorflow==2.7.0
pillow==8.4.0
opencv-python==4.5.4.60
